+++ src/css/search/search.css
@@ -4,6 +4,7 @@
   --color-form: rgba(5, 5, 5, 0.50);
 }
 
+
 .container {
   width: 100%; 
   padding-left: 20px;
@@ -11,6 +12,10 @@
   margin: 0 auto;
 }
 .gallery {
+    /* display: flex; */
+    /* width: 1140px; */
+    /* flex-wrap: wrap; */
+    /* gap: 12px; */
     padding: 0;
     list-style: none;
   }
@@ -32,6 +37,7 @@
   }
   .search-form-input {
     width: 335px;
+    /* height: 39px; */
   }
   .select-time {
     width: 160px;
@@ -73,6 +79,7 @@
   }
   .form-icon-search {
     text-align: center;
+    /* stroke: #9BB537; */
   }
   .saarch-form-btn:hover,
   .saarch-form-btn:focus {
@@ -84,6 +91,9 @@
   .form-icon-x {
     stroke: #9BB537;
   }
+  /* input:active .saarch-form-btn {
+    stroke: #9BB537;
+  } */
   .label-form {
     color: var(--color-form);
     font-size: 12px;
@@ -103,8 +113,9 @@
   
   .style-form option {
     border-radius: 14px;
+    border-color: brown;
     background: #FFF;
-    box-shadow: 0px 4px 36px 0px rgba(0, 0, 0, 0.02);
+    /* box-shadow: 0px 4px 36px 0px rgba(0, 0, 0, 0.02); */
     color: rgba(5, 5, 5, 0.30);
     font-family: Inter;
     font-size: 16px;
@@ -129,6 +140,8 @@
   .form-reset {
     display: flex;
     gap: 4px;
+    /* position: absolute;
+    right: 20px; */
     justify-content: flex-end;
     margin-bottom: 20px;
   }
@@ -144,6 +157,4 @@
 
   .gallery-oops {
     display: none;
-    justify-content: center;
   }
-
