+++ src/partials/search.html
@@ -4,61 +4,67 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Homework 11</title>
-    <link rel="stylesheet" href="./css/search.css">
   </head>
   <body>
-  <div class="content">
-    <form class="search-form" id="search-form">
-      <label class="label-form"> Search
-        <div class="search-form-input">
-          <button class="saarch-form-btn">
-            <svg class="form-icon-search" width="18" height="18">
-              <use href="./images/icons.svg#icon-search"></use>
-            </svg>
-          </button>
-          <input
+    <div class="content">
+      <form class="search-form" id="search-form">
+        <label class="label-form">
+          Search
+          <div class="search-form-input">
+            <button class="saarch-form-btn">
+              <svg class="form-icon-search" width="18" height="18">
+                <use href="./svg/symbol-defs.svg#iocn-search"></use>
+              </svg>
+            </button>
+            <input
               class="style-form saarch-form-txt"
               type="text"
               name="searchQuery"
               autocomplete="off"
               placeholder="Enter Text"
-          />
-          <button class="saarch-form-btn-close">
-            <svg class="form-icon-close" width="18" height="18">
-              <use href="./images/icons.svg#icon-x"></use>
-            </svg>
-          </button>
-        </div>
-      </label>
-      <label class="label-form">Time
+            />
+            <button class="saarch-form-btn-close">
+              <svg class="form-icon-close" width="18" height="18">
+                <use href="./svg/symbol-defs.svg#iocn-x"></use>
+              </svg>
+            </button>
+          </div>
+        </label>
+        <label class="label-form"
+          >Time
           <select class="style-form select-time" id="time" name="time">
             <option value disabled selected>Select</option>
           </select>
-      </label>
-      <label class="label-form">Area
+        </label>
+        <label class="label-form"
+          >Area
           <select class="style-form select-area" id="area" name="area">
             <option value disabled selected>Select</option>
           </select>
-      </label>
-      <label class="label-form">Ingredients
-          <select class="style-form select-ingredients" id="ingredients" name="ingredients">
+        </label>
+        <label class="label-form"
+          >Ingredients
+          <select
+            class="style-form select-ingredients"
+            id="ingredients"
+            name="ingredients"
+          >
             <option value disabled selected>Select</option>
           </select>
-      </label>
-    </form>
-    <div class="form-reset">
-      <svg class="form-icon-x" width="16" height="16">
-        <use href="./images/icons.svg#icon-x"></use>
-      </svg>
-      <p>Reset the filter</p>
+        </label>
+      </form>
+      <div class="form-reset">
+        <svg class="form-icon-x" width="16" height="16">
+          <use href="./svg/symbol-defs.svg#icon-x"></use>
+        </svg>
+        <p>Reset the filter</p>
+      </div>
+      
+      <div class="gallery"></div>
+      <li class="gallery-oops">
+        <img src="./images/oops-chef.jpg" alt="nothing found" />
+      </li>
     </div>
-    <div class="gallery">
-    </div>
-    <li class="gallery-oops">
-      <img src="./images/oops-chef.jpg" alt="nothing found">
-    </li>
-  </div>
-  
-  <script src="./js/search.js" type="module"></script>
+    <button type="button" class="load-more">Load more</button>
   </body>
 </html>
